requirements.txt
albucore==0.0.12
albumentations==1.4.11
annotated-types==0.7.0
colorlog==6.8.2
eval_type_backport==0.2.0
filelock==3.15.4
fsspec==2024.6.1
imageio==2.34.2
Jinja2==3.1.4
joblib==1.4.2
lazy_loader==0.4
lightning-utilities==0.11.6
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pillow==10.4.0
puts==0.0.11
pydantic==2.8.2
pydantic_core==2.20.1
PyYAML==6.0.1
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
sympy==1.13.1
threadpoolctl==3.5.0
tifffile==2024.7.24
tomli==2.0.1
torch==2.4.0
torchmetrics==1.4.0.post0
torchvision==0.19.0
tqdm==4.66.4
triton==3.0.0
typing_extensions==4.12.2
